requirements.txt
numpy
Pillow
scikit-learn
streamlit
reportlab
